==> 00 - Callings_preparation.ipynb <==
Importing the dataset and libraries.

import pandas as pd

callings=pd.read_csv("Data\sedec_chamados.csv",delimiter =';')

Renaming the column processo_numero to ID to match the other datasets, and while doing that dropping processo_origem.

callings = callings.rename(columns={'processo_numero':'ID'})



callings.drop(columns='processo_origem', axis=1, inplace=True)


Taking a look at the dataset.

callings.info()

callings.head()

Renaming the dataset for better understanding. This dataset relates to a calling made by the user and the why behind it, we are going to use the latitude and longitude values from this dataset, since it is the only one with that info. Besides that it informs us if there were victims at the scene, so we are also keeping that feature.

callings.drop(columns='ano', axis=1, inplace=True)
callings.drop(columns='mes', axis=1, inplace=True)
callings.drop(columns='solicitacao_data', axis=1, inplace=True)
callings.drop(columns='solicitacao_descricao', axis=1, inplace=True)
callings.drop(columns='solicitacao_regional', axis=1, inplace=True)
callings.drop(columns='solicitacao_bairro', axis=1, inplace=True)
callings.drop(columns='solicitacao_localidade', axis=1, inplace=True)
callings.drop(columns='solicitacao_endereco', axis=1, inplace=True)
callings.drop(columns='solicitacao_roteiro', axis=1, inplace=True)
callings.drop(columns='rpa_codigo', axis=1, inplace=True)
callings.drop(columns='rpa_nome', axis=1, inplace=True)
callings.drop(columns='solicitacao_microrregiao', axis=1, inplace=True)
callings.drop(columns='solicitacao_plantao', axis=1, inplace=True)
callings.drop(columns='processo_tipo', axis=1, inplace=True)
callings.drop(columns='processo_situacao', axis=1, inplace=True)
callings.drop(columns='solicitacao_hora', axis=1, inplace=True)
callings.head()

Transforming the columns with victims information into numerical values.

callings['solicitacao_vitimas'] = callings['solicitacao_vitimas'].astype('str')
callings['solicitacao_vitimas_fatais'] = callings['solicitacao_vitimas_fatais'].astype('str')


callings['solicitacao_vitimas']=[1 if sit == 'Sim' else 0 for sit in callings['solicitacao_vitimas']]
callings['solicitacao_vitimas_fatais']=[1 if sit == 'Sim' else 0 for sit in callings['solicitacao_vitimas_fatais']]

callings.head()

callings.drop(columns='solicitacao_origem_chamado', axis=1, inplace=True)
callings.drop(columns='processo_status', axis=1, inplace=True)
callings.drop(columns='processo_localizacao', axis=1, inplace=True)
callings.drop(columns='processo_data_conclusao', axis=1, inplace=True)
callings.head()

callings = callings.rename(columns={'solicitacao_vitimas':'victims',
                                 'solicitacao_vitimas_fatais':'deadly_victims'})
callings.head()

callings.info()

Checking for duplicate values.

callings = callings.drop_duplicates('ID')

callings.info()

Replacing the null values for 0.

callings.fillna(0,inplace=True)
callings.head()

Taking a look at the unique values and information of the dataset.

callings['latitude'].value_counts()

callings['victims'].value_counts()

callings['deadly_victims'].value_counts()

locations = callings.groupby(['latitude','longitude']).size().reset_index().rename(columns={0:'count'})
locations

number_locations = locations['count'].value_counts()
number_locations

len(number_locations)

callings.to_csv(path_or_buf='callings_prepared.csv')
