==> 00 - Canvas_preparation.ipynb <==
Importing the dataset and libraries.

import pandas as pd

canvas = pd.read_csv('Data/sedec_lonas.csv',sep=";")

Taking a look at the dataset.

canvas.head()

canvas.info()

Renaming the dataset for better understanding.
This dataset relates to if and when (the location will be given when we cross-reference the ID with the other datasets) a canvas was placed on the location: so in_place it means it was placed, justification means why it was placed, length the length of the canvas, and num_point the number of points it was placed at that specific location.

canvas = canvas.rename(columns={'ano':'year','mes':'month','processo_numero':'ID','colocacao_lona_situacao':'in_place',
                      'colocacao_lona_data':'date_of_placement','colocacao_lona_justificativa':'justification',
                      'colocacao_lona_metragem':'length','colocacao_lona_qtd_pontos':'num_points'})
canvas

Isolating the month and date values in the respective columns.

aux=[]
for i in range (0,len(canvas['month'])):
    aux = canvas['month'][i]
    canvas['month'][i] = aux[:2]
    if canvas['month'][i] == 'Nã':
        canvas['month'][i] = 0
    canvas['month'][i] = int(canvas['month'][i])

canvas['month'][:10]

canvas.head()

canvas['date_of_placement'].fillna(0,inplace=True)
aux=[]
for i in range (0,len(canvas['date_of_placement'])):
    if canvas['date_of_placement'][i] !=0:
        aux = canvas['date_of_placement'][i]
        canvas['date_of_placement'][i] = aux[:2]
        canvas['date_of_placement'][i] = int(canvas['date_of_placement'][i])

canvas['date_of_placement'][:10]

canvas.head()

Transforming the in_place column into numerical values.

canvas = pd.get_dummies(canvas, columns=['in_place'],drop_first=True)


canvas.head()

canvas = canvas.rename(columns={'in_place_Sim':'in_place'})
canvas.head()

After a look at the justification column, we detected that the information that we need is if a canvas was placed due to a sliding, so we'll isolate that keyword 'deslizamento'(sliding in portuguese) and transform into a 0, 1 column respectively.

import re
canvas['justification'] = canvas['justification'].astype(str)

text = []
for i in range(0, len(canvas['justification'])):
    review = re.sub('[^a-zA-Z]', ' ', canvas['justification'][i])
    review = review.lower()
    review = review.split()
    text.append(review)


            

canvas['justification'] = text

canvas.head()

canvas['justification'] = [1 if 'deslizamento' in justification else 0 for justification in canvas['justification']]
canvas.head()

Finally there are some duplicate ID entries, so we'll drop the ones that might have 0 values in the event of a sliding.

canvas = canvas.sort_values('justification', ascending=False).drop_duplicates('ID').sort_index()

After cross checking the date values with other datasets we decided to drop the date information on this one(see date_check for more info).

canvas = canvas.drop(columns=['year','month','date_of_placement'],axis=1)

Taking a look at the unique values of each column after the preparation.

canvas['length'].value_counts()

canvas['num_points'].value_counts()

canvas['in_place'].value_counts()

canvas['justification'].value_counts()

canvas.head()

canvas.to_csv(path_or_buf='canvas_prepared.csv')
