==> 00 - Events_preparation.ipynb <==
Importing the dataset and libraries.

import pandas as pd


events = pd.read_csv('Data/sedec_tipo_ocorrencias.csv',sep=';')

Taking a look at the dataset.

events.info()

events.head()

events['processo_ocorrencia'].value_counts()

Renaming the dataset for better understanding. This dataset relates to what happened on a given ID (the location and date will be given when we cross-reference the ID with the other datasets).

events = events.rename(columns={'processo_numero':'ID','processo_ocorrencia':'event'})

events.head()

After a look at the event column, we detected that the information that we need is if sliding took place, so we'll isolate that keyword 'deslizamento'(sliding in portuguese) and transform into a 0, 1 column respectively.

import re


text = []
for i in range(0, len(events['event'])):
    review = re.sub('[^a-zA-Z]', ' ', events['event'][i])
    review = review.lower()
    review = review.split()
    text.append(review)



events['event'] = text
events['event'] = [1 if 'deslizamentos' in event else 0 for event in events['event']]
events.head()

events.event.value_counts()

Finally there are some duplicate ID entries, so we'll drop the ones that might have 0 values in the event of a sliding.

events = events.sort_values('event', ascending=False).drop_duplicates('ID').sort_index()

events.to_csv(path_or_buf='events_prepared.csv')
