==> 00 - Inspections_preparation.ipynb <==
Importing the dataset and libraries.

import pandas as pd
import numpy as np

inspections = pd.read_csv('Data/sedec_vistorias.csv',sep=";")

Taking a look at the dataset.

inspections.info()

inspections.head()

inspections['vistoria_risco'].value_counts()

inspections['vistoria_localidade'].value_counts()

inspections['ano'].value_counts()

We are going to drop all the columns that don't relate to the info that we need, all of these relate to the inspection process and log so these features aren't related to slidings. Additionally, we are dropping the date and locations columns since we'll be extracting these values from a different dataset when cross-referencing the ID.

inspections = inspections.drop(columns=['ano','mes','vistoria_data','avaliador',
                        'vistoria_localidade','vistoria_rpa_codigo','vistoria_setor',
                        'vistoria_microrregiao'],axis=1)

inspections.head()

Renaming the dataset for better understanding. This dataset relates to an inspection conducted on that place, the feature that relates to our problem is the risk evaluated by the inspector at the scene.

inspections = inspections.rename(columns={'vistoria_risco':'risk','processo_numero':'ID'})
inspections.head()

We will transform the risk column into:
4- Very High risk.
3- High.
2- Medium.
1- Low.
0- Not informed.

mapping = {'Não informado': 0, 'R1 Baixo': 1,'R2 Médio': 2,'R3 Alto': 3,'R4 Muito Alto': 4 }

inspections = inspections.applymap(lambda s: mapping.get(s) if s in mapping else s)
inspections.head()

Finally there are some duplicate ID entries, so we'll drop the ones that might have 0 values for risk assessment.

inspections = inspections.sort_values('risk', ascending=False).drop_duplicates('ID').sort_index()

Taking a look at the unique values of this dataset.

inspections['risk'].value_counts()

inspections.to_csv(path_or_buf='inspections_prepared.csv')
