==> 00 - Rain_preparation.ipynb <==
Importing the dataset and libraries.

import pandas as pd

rain = pd.read_csv('Data/rain_Recife_2012_2019.csv')

Taking a look at the dataset.

rain.head()

rain.info()

Separating the year from the month values and assign each one to his own column.

rain = rain.rename(columns={'Mês/Ano':'month_year','Posto':'Station'})

rain.head()

rain['year']= rain['month_year'].copy()

rain.head()

aux=[]
for i in range (0,len(rain['month_year'])):
    if rain['month_year'][i] !=0:
        aux = rain['month_year'][i]
        rain['month_year'][i] = aux[:3]

rain['month_year'].value_counts()

aux=[]
for i in range (0,len(rain['year'])):
    if rain['year'][i] !=0:
        aux = rain['year'][i]
        rain['year'][i] = aux[-4:]
        rain['year'][i] = int(rain['year'][i])

rain['year'][:10]

Transforming the months into numerical values.

mapping = {'jan': 1, 'fev': 2,'mar': 3,'abr': 4,'mai': 5,'jun':6, 'jul':7,'ago':8,'set':9,'out':10,'nov':11,'dez':12}

rain = rain.applymap(lambda s: mapping.get(s) if s in mapping else s)
rain.head()

rain['month_year'].value_counts()

rain = rain.rename(columns={'month_year':'month'})

This dataset is the rain registered at that station on that day so in order to merge with the other dataset, there is a need to reorder how the information is shown. The day needs to be in a single column matched with it's respective value. While doing so we will clear the null values.

column_names = list(rain.columns.values)
column_names = column_names[3:-2]
temporary_row= [[0,0,0,0,0,0]]
rain_transformed = pd.DataFrame(temporary_row,columns=['Código','Station','month','year','rain','day'])
temporary=[]
for i in column_names:
    temporary = rain[['Código','Station','month','year',i]]
    temporary = temporary.rename(columns={i:'rain'})
    temporary = temporary[temporary.rain != '-']
    day = [i for n in range(0,len(temporary['rain']))]
    temporary['day']= day
    rain_transformed = pd.concat([rain_transformed,temporary], axis=0, join='outer', ignore_index=False)

rain_transformed['day'] = rain_transformed.day.astype(int)
rain_transformed = rain_transformed.drop([0])
rain_transformed.head()


Taking a look at the unique values.

rain_transformed.day.value_counts().sort_index()

rain_transformed.info()

rain_transformed.Station.value_counts()

Adding the latitude and longitude of the stations.

rain_transformed['latitude']=0
rain_transformed['longitude']=0
rain_transformed.head()

rain_transformed.loc[rain_transformed['Station']=='Recife (Lamepe/Itep)','latitude'] = -8.0592
rain_transformed.loc[rain_transformed['Station']=='Recife (Lamepe/Itep)','longitude'] = -34.9519
rain_transformed.loc[rain_transformed['Station']=='Recife (Santo Amaro)','latitude'] = -8.0503
rain_transformed.loc[rain_transformed['Station']=='Recife (Santo Amaro)','longitude'] = -34.8842
rain_transformed.loc[rain_transformed['Station']=='Recife (Várzea)','latitude'] = -8.036
rain_transformed.loc[rain_transformed['Station']=='Recife (Várzea)','longitude'] = -34.98
rain_transformed.loc[rain_transformed['Station']=='Recife - PCD','latitude'] = -8.0644
rain_transformed.loc[rain_transformed['Station']=='Recife - PCD','longitude'] = -34.9247
rain_transformed.loc[rain_transformed['Station']=='Recife (Alto da Brasileira)','latitude'] = -8.0014
rain_transformed.loc[rain_transformed['Station']=='Recife (Alto da Brasileira)','longitude'] = -34.9353
rain_transformed.loc[rain_transformed['Station']=='Recife (Codecipe / Santo Amaro)','latitude'] = -8.0442
rain_transformed.loc[rain_transformed['Station']=='Recife (Codecipe / Santo Amaro)','longitude'] = -34.8747
rain_transformed[:10]

Transforming the rain values into float type values.

rain_transformed['rain']= rain_transformed['rain'].str.replace(',','.')
rain_transformed['rain']= rain_transformed['rain'].astype(float)
rain_transformed.head()

Converting the date information to a pandas time series.

import datetime

rain_transformed['Date'] = pd.to_datetime(rain_transformed[['year', 'month', 'day']])
rain_transformed.head()

Droping the unnecessary columns.

rain_transformed = rain_transformed.drop(columns=['year','month','day','Código'],axis=1)
rain_transformed.head()

rain_transformed.to_csv('rain_prepared.csv')
