==> 00 - Solicitations_preparation.ipynb <==
Importing the dataset and libraries.

import pandas as pd

solicitations=pd.read_csv("Data\sedec_solicitacoes.csv",delimiter =';')

Renaming the column processo_numero to ID to match the other datasets, and while doing that dropping processo_origem.

solicitations = solicitations.rename(columns={'processo_numero':'ID'})


solicitations.drop(columns='processo_origem', axis=1, inplace=True)

Taking a look at the dataset.

solicitations.info()

solicitations.head()

In the solicitacao_descricao column there is information about if a sliding took place. Additionally the processo_situacao column says if the process is considered complete or not. So we are going to isolate the keywords deslizamento and completo and transform into 0/1 values.

solicitations['processo_situacao'] = solicitations['processo_situacao'].astype('str')
solicitations['solicitacao_descricao'] = solicitations['solicitacao_descricao'].astype('str')
solicitations['solicitacao_descricao'].str.lower()

solicitations['solicitacao_descricao']=[1 if 'deslizamento' in review else 0 for review in solicitations['solicitacao_descricao']]
solicitations['processo_situacao']=[1 if sit == 'completo' else 0 for sit in solicitations['processo_situacao']]

solicitations

Dropping the columns that have no relevant information.

solicitations.drop(columns='solicitacao_hora', axis=1, inplace=True)
solicitations.drop(columns='solicitacao_plantao', axis=1, inplace=True)
solicitations.drop(columns='solicitacao_localidade', axis=1, inplace=True)
solicitations.drop(columns='rpa_nome', axis=1, inplace=True)
solicitations.drop(columns='rpa_codigo', axis=1, inplace=True)
solicitations.drop(columns='solicitacao_regional', axis=1, inplace=True)
solicitations.drop(columns='solicitacao_microrregiao', axis=1, inplace=True)
solicitations.drop(columns='solicitacao_endereco', axis=1, inplace=True)
solicitations.drop(columns='solicitacao_roteiro', axis=1, inplace=True)
solicitations.drop(columns='processo_solicitacao', axis=1, inplace=True)
solicitations.head()

solicitations.info()

Renaming the columns for better understading.

solicitations = solicitations.rename(columns={'ano':'year','mes':'month','solicitacao_data':'day',
                                 'solicitacao_descricao':'description','solicitacao_bairro':'neighbourhood',
                                 'processo_situacao':'situation'})
solicitations.head()

Dropping the duplicate values.

solicitations = solicitations.sort_values('description', ascending=False).drop_duplicates('ID').sort_index()
solicitations.info()

Isolating the month and day values in the respective columns.

aux=[]
for i in range (0,len(solicitations['month'])):
    aux = solicitations.loc[solicitations.index[i],'month']
    solicitations.loc[solicitations.index[i],'month'] = aux[:2]
    solicitations.loc[solicitations.index[i],'month'] = int(solicitations.loc[solicitations.index[i],'month'])

solicitations['month'][:10]

aux=[]
for i in range (0,len(solicitations['day'])):
    aux = solicitations.loc[solicitations.index[i],'day']
    solicitations.loc[solicitations.index[i],'day'] = aux[-2:]
    
solicitations['day'][:10]

solicitations.head()

solicitations['day'] = solicitations['day'].astype(int)
solicitations.head()

There is still one column with no relevant information. We are keeping the neighbourhood column because there are no null values in this one, so we can cross-reference that with the latitude/longitude values and complete the data about the locale.

solicitations = solicitations.drop(columns='processo_tipo',axis=1)
solicitations.head()

Taking a look at the unique values and information of the dataset.

solicitations['description'].value_counts()

solicitations['situation'].value_counts()

solicitations['neighbourhood'].value_counts()

locations = solicitations.groupby(['neighbourhood']).size().reset_index().rename(columns={0:'count'})
locations[70:90]

number_locations = locations['count'].value_counts()
number_locations

len(number_locations)

Converting the date information to a pandas time series.

import datetime

solicitations['Date'] = pd.to_datetime(solicitations[['year', 'month', 'day']])
solicitations.head()

Droping the unnecessary date columns.

solicitations = solicitations.drop(columns=['year','month','day'],axis=1)
solicitations.head()

The analysis ends in december 2019 so we need to remove the first row.

solicitations = solicitations.drop([0])
solicitations.head()

Transforming the neighbourhood data into numerical values, we will just transform it into a sequential integer list. Additionally there are some rows which do not have information about the location ('Não informado') so we need to remove those rows.

locations= locations[locations['neighbourhood']!='Não informado']
locations['list']=range(1,120)
locations = locations.drop(columns='count',axis=1)
locations.head()

solicitations['location']= 0
solicitations = solicitations[solicitations['neighbourhood']!='Não informado']
solicitations.head()

solicitations.info()

for n in range (0,len(solicitations)):
    b= locations.loc[locations['neighbourhood']==solicitations.neighbourhood.iloc[n]]
    solicitations.location.iloc[n]= b.list.iloc[0]

solicitations.head()

Dropping the neighbourhood column since there is no need for it anymore.

solicitations = solicitations.drop(columns='neighbourhood',axis=1)
solicitations.head()

solicitations.to_csv(path_or_buf='solicitations_prepared.csv')
