==> 01 - Slidings_merging_Datasets.ipynb <==
Importing the datasets and libraries.

import pandas as pd

canvas = pd.read_csv('Data/canvas_prepared.csv')
events = pd.read_csv('Data/events_prepared.csv')
inspections = pd.read_csv('Data/inspections_prepared.csv')
solicitations = pd.read_csv('Data/solicitations_prepared.csv')
callings = pd.read_csv('Data/callings_prepared.csv')
canvas = canvas.drop(columns='Unnamed: 0',axis=1)
inspections = inspections.drop(columns='Unnamed: 0',axis=1)
events = events.drop(columns='Unnamed: 0',axis=1)
solicitations = solicitations.drop(columns='Unnamed: 0',axis=1)
callings = callings.drop(columns='Unnamed: 0',axis=1)

Taking a quick look at the datasets and the unique values at the most critical columns, we are doing this to ensure that we don't lose or duplicate information after a merge.

canvas.info()

canvas['justification'].value_counts()

events.info()

events['event'].value_counts()

inspections.info()

callings.info()

solicitations.info()

solicitations['description'].value_counts()

Merging the datasets.

slidings = solicitations.set_index('ID').join(inspections.set_index('ID')).join(events.set_index('ID')).join(callings.set_index('ID')).join(canvas.set_index('ID'))

slidings.head()

slidings.info()

slidings['risk'].value_counts()

slidings['event'].value_counts()

slidings['justification'].value_counts()

slidings['description'].value_counts()

Putting back the 'ID' as a column instead of an index.

slidings.reset_index(level=0,inplace=True)

slidings.head()

Replacing the null values for zeros.

slidings.fillna(0,inplace=True)

slidings.head()

Transforming every 0,1 column to integer type variables.

slidings['risk'] = slidings['risk'].astype(int)
slidings['event'] = slidings['event'].astype(int)
slidings['victims'] = slidings['victims'].astype(int)
slidings['deadly_victims'] = slidings['deadly_victims'].astype(int)
slidings['justification'] = slidings['justification'].astype(int)
slidings['length'] = slidings['length'].astype(int)
slidings['num_points'] = slidings['num_points'].astype(int)
slidings['in_place'] = slidings['in_place'].astype(int)

slidings.head()

slidings.info()

From this point on we no longer need the ID column so we will drop it.

slidings = slidings.drop(columns='ID',axis=1)
slidings.head()

There are 3 columns with the same information (0- no sliding, 1- sliding), which are situation, event and justification. So we are going to add them, take a quick look at how many slidings are reported at more than one dataset (values higher than 1) and then finally converting to 0/1 again.

slidings['sliding'] = slidings.apply(lambda row: row.situation + row.event + row.justification , axis = 1) 
slidings.sliding.value_counts()

slidings.sliding.replace(2,1,inplace=True)
slidings.sliding.value_counts()

Since the sliding information is concentrade in the new sliding column, we can drop the previous ones.

slidings = slidings.drop(columns=['description','event','justification'],axis=1)
slidings.head()

slidings.to_csv(path_or_buf='slidings_merged.csv')
