==> 01.1 - Date_check.ipynb <==
Based on the naming of the dataset there were some assumptions as to what each date represented, so we decided to check that.

import pandas as pd


solicitations = pd.read_csv('Data/solicitations_prepared.csv')
solicitations = solicitations.drop(columns='Unnamed: 0',axis=1)
canvas = pd.read_csv('Data/canvas_prepared_wdate.csv')
canvas = canvas.drop(columns='Unnamed: 0',axis=1)

First of all, the canvas dataset date lead us to believe, that it represented when a canvas was placed, and the solicitation date, the date where the process begun. So we will isolate some date values per ID on both datasets and compare it.

canvas.loc[canvas['ID'] == 8002630913]

solicitations.loc[solicitations['ID'] == 8002630913]

canvas.loc[canvas['ID'] == 8002108013]

solicitations.loc[solicitations['ID'] == 8002108013]

We can see that the dates on the solicitations are after the dates on the canvas dataset, which means that our initial assumption is wrong, Concluding, we will use the date on the solicitations dataset since it is the only one which is complete (there are zero null values for date there) and work with that date to build our model.
