==> 02 - Slidings_Rain_merge.ipynb <==
Importing the datasets and libraries.

import pandas as pd
import numpy as np

slidings = pd.read_csv('Data/slidings_merged.csv')
rain = pd.read_csv('Data/rain_prepared.csv')

Taking a look at the datasets.

slidings = slidings.drop(columns='Unnamed: 0',axis=1)
slidings.head()

rain = rain.drop(columns='Unnamed: 0',axis=1)
rain.head()

rain.info()

Creating the rain column to fill with the value from the nearest station on that day. Those stations with no latitude/longitude values, will be filled with the average of the stations on the respective day.

slidings['Rain']=0

from geopy.distance import geodesic
for n in range (0,len(slidings)):
    b= rain.loc[rain['Date']==slidings.Date.iloc[n]]
    p1 = [slidings['latitude'][n],slidings['longitude'][n]]
    point =[]
    distance = []
    if slidings.latitude[n] == 0:
        slidings.Rain.iloc[n]= b['rain'].sum()/len(b['rain'])
    else:
            for i in range (0,len(b)):
                point = [b.latitude.iloc[i],b.longitude.iloc[i]]
                distance.append(geodesic(p1, point).meters)
            slidings.Rain.iloc[n]=b.rain.iloc[distance.index(min(distance))]

            
slidings['Rain']= np.round(slidings['Rain'],decimals=2)

slidings.Rain.value_counts()

The date, latitude and longitude columns are no longer needed so we will drop them. We will use the location list as the feature containing the location information, since it is the only that it is complete.

slidings = slidings.drop(columns=['latitude','longitude','Date'],axis=1)
slidings.head()

Quick check for any null values. Since it could happen that no station had recorded values for rain on the day that we looked up.

slidings.isnull().any()

slidings[slidings.isnull().any(axis=1)]['Rain']

Removing the null value.

slidings = slidings.drop([3824])
slidings.isnull().any()

slidings.to_csv(path_or_buf='slidings_training.csv')
